Linux: Only add icons and themes if they don't exist

+++ install/linux/icons.sh
@@ -1,5 +1,7 @@
-git clone https://github.com/vinceliuice/Tela-icon-theme.git
-cd Tela-icon-theme
-./install.sh -a
-cd ..
-rm -rf Tela-icon-theme
+if [ ! -d "${HOME}/.local/share/icons/Tela" ]; then
+    git clone https://github.com/vinceliuice/Tela-icon-theme.git
+    cd Tela-icon-theme
+    ./install.sh -a
+    cd ..
+    rm -rf Tela-icon-theme
+fi
