+++ install/linux/themes.sh
@@ -1,5 +1,7 @@
-git clone https://github.com/vinceliuice/Orchis-theme.git
-cd Orchis-theme
-./install.sh
-cd ..
-rm -rf Orchis-theme
+if [ ! -d "${HOME}/.themes/Orchis" ]; then
+    git clone https://github.com/vinceliuice/Orchis-theme.git
+    cd Orchis-theme
+    ./install.sh
+    cd ..
+    rm -rf Orchis-theme
+fi
